% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/at_animate.R
\name{at_animate}
\alias{at_animate}
\title{Animated Atlantis Plots}
\usage{
at_animate(atl, interval = 0.2, codes)
}
\arguments{
\item{atl}{Atlantis output returned from \code{read_atlantis()}}

\item{interval}{Speed of animination (unit in seconds)}

\item{codes}{Vector of Atlantis functional group codes to create GIFs}
}
\description{
Create animated Atlantis plots of all the tracers
}
\details{
This function will create GIFs of all Atlantis tracers by default. WARNING: This can take a while. It is probably much, much better to specify certain groups!
}

