% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gen_salt.R
\name{gen_salt}
\alias{gen_salt}
\title{Generate salinity forcing data for Atlantis}
\usage{
gen_salt(b, z, time, salt, start_time, fill_value = 35, model_name, bgm_file,
  gen_nc = FALSE, keep_cdf = TRUE)
}
\arguments{
\item{b}{Number of boxes}

\item{z}{Number of vertical boxes (i.e. layers)}

\item{time}{A vector of time points (in seconds) between consecutive measurements}

\item{salt}{A \code{b} by \code{z} by \code{time} data.frame or matrix with each entry corresponding to a salinity measurement at a specific time.}

\item{start_time}{String variable of the format YYYY-MM-DD HH:MM:SS +0}

\item{fill_value}{Fill value for missing data. Defaults to 35.}

\item{model_name}{Name of the model.}

\item{bgm_file}{Name of the bgm file. Include the file extension!}

\item{gen_nc}{Generate the nc binary? Defaults to FALSE and required netcdf-bin to be installed.}

\item{keep_cdf}{Keep the readable cdf file? Defaults to TRUE.}
}
\description{
This function generates the salinity forcing file for Atlantis.
}
\details{
This function generates the salinity forcing data for Atlantis in the ncdf4 file format and generates a binary. This function can compress the resultant cdf file if \code{gen_nc = TRUE} is set and can clean up after itself if \code{keep_cdf = FALSE is set}.
}
\examples{
salt <- matrix(c(35.1345, 35.1319, 35.1598, 35.1781, 35.1754, 35.1649, 35.1345, 35.1241, 35.1951, 35.2213, 35.2308, 35.2257, 35.2146, 35.1241),nrow = 2)
b <- 1
z <- 7
bgm_file <- "atlantis_test.bgm"
model_name <- "atlantis"
start_time = "2014-01-01 00:00:00 +0"
time <- c(12,24)
gen_salt(b=b,z=z,time=time, salt = salt, start_time = start_time, model_name= model_name, bgm_file=bgm_file, gen_nc=TRUE)
}
\seealso{
\code{\link{gen_temp}},\code{\link{gen_init},\code{\link{dummy_hydro}}}
}

