% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_at.R
\name{plot_at}
\alias{plot_at}
\title{Atlantis Plots}
\usage{
plot_at(atl, type, use.log = TRUE, save.pdf = FALSE, funcode = NULL)
}
\arguments{
\item{atl}{Atlantis output returned from \code{read_atlantis()}}

\item{type}{\code{ssb}, \code{relbio}, or \code{all}. See details.}

\item{use.log}{Should a log transformation be used?}

\item{save.pdf}{Should a PDF be saved?}

\item{funcode}{Which functional groups to plot? Vector containing the Atlantis codes.}
}
\description{
Provides a suite of Atlantis output plots
}
\details{
\code{atplot} provides a suite of plot for Atlantis plots. Presently, it is able to plot spawning stock biomass (\code{type = "ssb"}) or relative biomass (\code{type = "relbio"}). Both types of plots are available simulataneously by specifying \code{type = "all"}. However, no plots will be printed to the \code{R} console but instead will be saved directly as PDFs.
}

