% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_atprod.R
\name{plot_atprod}
\alias{plot_atprod}
\title{Production plots from Atlantis}
\usage{
plot_atprod(data, type, grpname, agestruct = TRUE, save = FALSE)
}
\arguments{
\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. See below}

\item{type}{Either total or average}
}
\description{
Produces production plots from Eat, Growth, Spawning Size, and Ratios from the read_prod() function.
}

