% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_netcdf.R, R/read_prod.R
\name{read_prod}
\alias{read_prod}
\title{Read in netCDF production file for vertebrates and save it as a data.table}
\usage{
read_prod(file, grpname, startyear, agestruct = TRUE, ratios = FALSE)

read_prod(file, grpname, startyear, agestruct = TRUE, ratios = FALSE)
}
\arguments{
\item{grpname}{Which functional groups to plot? Vector containing the Atlantis functional group names. Omit for ratio indicators.}

\item{startyear}{Starting year of the model}

\item{agestruct}{Is the functional group age structured?}

\item{ratios}{Extract other ratio biomass indicators? Default: FALSE}

\item{grpname}{Which functional groups to plot? Vector containing the Atlantis functional group names. Omit for ratio indicators.}

\item{ratios}{Extract other ratio biomass indicators? Default: FALSE}

\item{startyear}{Starting year of the model}
}
\description{
Read in netCDF production file for vertebrates and save it as a data.table

Read in netCDF production file for vertebrates and save it as a data.table
}

