\name{rlantis-package}
\alias{rlantis-package}
\alias{rlantis}
\docType{package}
\title{
Atlantis Helper Tool for R
}
\description{
Set of helper functions for creating, manipulating, and viewing Atlantis files from R. 
}
\details{
At present the package, generates the initial conditions file for Atlantis (\code{\link{gen_init}}); generate dummy hydrography data for Atlantis (\code{\link{gen_hydro}}); and includes a data set (\code{\link{required_init}}) of required tracer, phys, and epibenthos variables.

For a complete list of functions, use library(help = "rlantis")
}