% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ssb_plot.R
\name{ssb_plot}
\alias{ssb_plot}
\title{Plot SSB biomass}
\usage{
ssb_plot(atl, use.log = TRUE, save.pdf = FALSE, funcode = NULL, ...)
}
\arguments{
\item{atl}{Atlantis output returned from read_atlantis()}

\item{use.log}{Should a log transformation be used?}

\item{save.pdf}{Should a PDF be saved?}

\item{funcode}{Which functional groups to plot? Vector containing the Atlantis codes}
}
\description{
Plot SSB biomass
}

