\docType{data}
\name{biol_template}
\alias{biol_template}
\title{Biological parameters template file}
\usage{
biol_template
}
\description{
This data set uses a skeleton biological parameters file
and inserts data from the user and from Fishbase.
}
\seealso{
\code{\link{bio_dem}}. \code{\link{bio_com}}
}

