\docType{package}
\name{rlantis-package}
\alias{rlantis}
\alias{rlantis-package}
\title{Helper tools for creating, manipulating, and viewing Atlantis input and output files}
\description{
Helper tools for creating, manipulating, and viewing
Atlantis input and output files
}
\details{
rlantis is a set of helper functions for creating,
manipulating, and viewing Atlantis files from R.
}
\author{
\email{cddesjardins@gmail.com}
}
\keyword{package}

